packaged server specific endpoints
