+++ src/main/java/site/bmraubo/echoServerEndpoints/HealthCheck.java
@@ -1,20 +1,21 @@
+package site.bmraubo.echoServerEndpoints;
+
 import site.bmraubo.HTTPServer.Endpoint;
 import site.bmraubo.HTTPServer.Request;
 import site.bmraubo.HTTPServer.Response;
 import site.bmraubo.HTTPServer.ResponseBuilder;
 
-public class Redirect implements Endpoint {
-    String[] allowedMethods = {"GET"};
-    String contentLocation = "http://127.0.0.1:5000/simple_get";
+public class HealthCheck implements Endpoint {
+    String contentType = "text/html;charset=utf-8";
+    String responseBody = "<html><body><p><strong>Status:</strong> pass</p></body></html>";
 
     @Override
     public Response prepareResponse(Request request) {
         ResponseBuilder responseBuilder = new ResponseBuilder();
         Response response = new Response(responseBuilder);
-        System.out.println("Redirecting");
-        responseBuilder.setStatusCode(301);
-        responseBuilder.setHeader("Location", contentLocation);
-        responseBuilder.setResponseBody("");
+        responseBuilder.setStatusCode(200);
+        responseBuilder.setHeader("Content-Type", contentType);
+        responseBuilder.setResponseBody(responseBody);
         return response;
     }
 }
