+++ src/main/java/site/bmraubo/echoServerEndpoints/MethodOptions.java
@@ -1,3 +1,5 @@
+package site.bmraubo.echoServerEndpoints;
+
 import site.bmraubo.HTTPServer.Endpoint;
 import site.bmraubo.HTTPServer.Request;
 import site.bmraubo.HTTPServer.Response;
