+++ src/main/java/site/bmraubo/echoServerEndpoints/Redirect.java
@@ -1,20 +1,22 @@
+package site.bmraubo.echoServerEndpoints;
+
 import site.bmraubo.HTTPServer.Endpoint;
 import site.bmraubo.HTTPServer.Request;
 import site.bmraubo.HTTPServer.Response;
 import site.bmraubo.HTTPServer.ResponseBuilder;
 
-public class XMLResponse implements Endpoint {
-    String contentType = "application/xml;charset=utf-8";
-    String responseBody = "<note><body>XML Response</body></note>";
+public class Redirect implements Endpoint {
+    String[] allowedMethods = {"GET"};
+    String contentLocation = "http://127.0.0.1:5000/simple_get";
 
     @Override
     public Response prepareResponse(Request request) {
         ResponseBuilder responseBuilder = new ResponseBuilder();
         Response response = new Response(responseBuilder);
-        responseBuilder.setStatusCode(200);
-        responseBuilder.setHeader("Content-Type", contentType);
-        responseBuilder.setResponseBody(responseBody);
+        System.out.println("Redirecting");
+        responseBuilder.setStatusCode(301);
+        responseBuilder.setHeader("Location", contentLocation);
+        responseBuilder.setResponseBody("");
         return response;
     }
-
 }
