+++ src/main/java/site/bmraubo/echoServerEndpoints/doggo.png
