+++ src/main/java/site/bmraubo/echoServerEndpoints/kitteh.jpg
